+++ source/common/config/utility.cc
@@ -32,8 +32,9 @@ void Utility::translateApiConfigSource(const std::string& cluster, uint32_t refr
     envoy::api::v2::core::GrpcService* grpc_service = api_config_source.add_grpc_services();
     grpc_service->mutable_envoy_grpc()->set_cluster_name(cluster);
   } else {
-    if (api_type == ApiType::get().RestLegacy) {
-      api_config_source.set_api_type(envoy::api::v2::core::ApiConfigSource::REST_LEGACY);
+    if (api_type == ApiType::get().UnsupportedRestLegacy) {
+      api_config_source.set_api_type(
+          envoy::api::v2::core::ApiConfigSource::UNSUPPORTED_REST_LEGACY);
     } else if (api_type == ApiType::get().Rest) {
       api_config_source.set_api_type(envoy::api::v2::core::ApiConfigSource::REST);
     }
@@ -144,7 +145,7 @@ void Utility::checkApiConfigSourceSubscriptionBackingCluster(
       (api_config_source.api_type() == envoy::api::v2::core::ApiConfigSource::GRPC);
 
   if (!api_config_source.cluster_names().empty()) {
-    // All API configs of type REST and REST_LEGACY should have cluster names.
+    // All API configs of type REST and UNSUPPORTED_REST_LEGACY should have cluster names.
     // Additionally, some gRPC API configs might have a cluster name set instead
     // of an envoy gRPC.
     Utility::validateClusterName(clusters, api_config_source.cluster_names()[0]);
@@ -180,7 +181,7 @@ void Utility::translateCdsConfig(const Json::Object& json_config,
                                  envoy::api::v2::core::ConfigSource& cds_config) {
   translateApiConfigSource(json_config.getObject("cluster")->getString("name"),
                            json_config.getInteger("refresh_delay_ms", 30000),
-                           json_config.getString("api_type", ApiType::get().RestLegacy),
+                           json_config.getString("api_type", ApiType::get().UnsupportedRestLegacy),
                            *cds_config.mutable_api_config_source());
 }
 
@@ -196,7 +197,7 @@ void Utility::translateRdsConfig(
 
   translateApiConfigSource(json_rds.getString("cluster"),
                            json_rds.getInteger("refresh_delay_ms", 30000),
-                           json_rds.getString("api_type", ApiType::get().RestLegacy),
+                           json_rds.getString("api_type", ApiType::get().UnsupportedRestLegacy),
                            *rds.mutable_config_source()->mutable_api_config_source());
 }
 
@@ -205,7 +206,7 @@ void Utility::translateLdsConfig(const Json::Object& json_lds,
   json_lds.validateSchema(Json::Schema::LDS_CONFIG_SCHEMA);
   translateApiConfigSource(json_lds.getString("cluster"),
                            json_lds.getInteger("refresh_delay_ms", 30000),
-                           json_lds.getString("api_type", ApiType::get().RestLegacy),
+                           json_lds.getString("api_type", ApiType::get().UnsupportedRestLegacy),
                            *lds_config.mutable_api_config_source());
 }